Add files via upload
"Love Calculator" is a Python project utilizing the tkinter module to create a graphical user interface (GUI) application. It allows users to calculate and analyze their love compatibility with someone else based on inputted information. This project provides a fun and interactive way to explore the dynamics of relationships using programming.